pyproject.toml
[project]
name = "trading-bot"
version = "0.1.0"
description = "Alpaca paper trading bot with Streamlit dashboard"
requires-python = ">=3.12"
dependencies = [
    "alpaca-py>=0.30.0",
    "pandas>=2.2.0",
    "streamlit>=1.40.0",
    "ta>=0.11.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "pytz>=2026.1.post1",
]
